Added check for img tags without src attributes
If an image tag is defined without a src attribute, it can cause some
browsers to make requests to "/", which can have dramatic performance
implications.  Put severity to RED (the only set of rules currently at
this high state) because of the big performance ding, difficulty in
quickly finding the issue and how easy it is for developers to
accidentally pass a null string when building markup.