Changing the structure README file
First, it's using the markdown syntax.

The second is to add the title of the project (to make it clearer that this is). After that, two unnecessary headers were deleted (because the text is small, they are not particularly useful).

Finally, the transfer of the text about the license to the end of the file (again for greater readability).